python-llm-service/requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
fastapi==0.115.0
uvicorn==0.32.0
